Let commands declaring their parser
