+++ deepomatic/cli/cmds/parser_helpers.py
@@ -1,0 +1,18 @@
+# Common parsers functions used by various commands
+
+COMMON_GROUPS = {
+    'input', 'output'
+}
+
+def add_common_cmd_group(parser, label):
+    assert label in COMMON_GROUPS
+    return parser.add_argument_group('{} arguments'.format(label.lower()))
+
+
+def add_recursive_param(parser):
+    parser.add_argument('-R', '--recursive', dest='recursive', action='store_true',
+                        help='If a directory input is used, goes through all files in subdirectories.')
+
+def add_verbose_param(parser):
+    parser.add_argument('--verbose', dest='verbose', action='store_true',
+                        help='Increase output verbosity.')
